Updated openstack/openstack Project: openstack-infra/elastic-recheck  53431da47c2859708e199fa329172b421a29358e
Add e-r query for bug/1646779

Add an E-R query for bug/1646779. The error message in n-cpu is about
libvirt not reachable, but the cause of that is that libvirt is killed
by the kernel because of either a general protection or a stack segment
trap. Both cases happen close to a virtlogd error message:

"End of file while reading data: Input/output error"

However the frequency of that is 3/4 times more than that of the result
of this query, which suggests that that error message is not always
related to libvirt getting killed, even if the other way around seems
to be the case.

Change-Id: I94a96469c21805b266beeee8c4863001886ab549
Related-Bug: #1646779