Updated openstack/openstack Project: openstack-dev/devstack  b6a7461a839ae08d852b33820a713d34d5a89948
Always add the service role to service users

When creating a service user we allow the user to be created with a
different role. Currently in auth_token middleware we want to check that
the service token is specified with the service role so we should always
add the service role and optionally add additional roles.

Change-Id: Ie954a679674b4795079b539ebc8d4d2dcbd7dacc