Updated openstack/openstack Project: openstack/nova  784db5a1d73a304562404dd895f15da8963528aa
Fix error if free_disk_gb is None in CellStateManager

The field 'free_disk_gb' in Objects.ComputeNode can be nullable.
In CellStateManager._get_compute_hosts, we compute free_disk_mb
by 'free_disk_gb * 1024'. If the free_disk_gb is None, it will raise
a TypeError exception because None type can't multiply by an integer.
We should check this value, and if it's None, we just let it return
zero.

Closes-Bug: #1648983

Change-Id: I558d544ff838807c1d605cb5ed8d0e0fa97ab00e