Updated openstack/openstack Project: openstack-infra/nodepool  759862b840b965cc8fd563a16dbc050c8dfa0cfa
Expose the ability to build all infra diskimages

Add the ability for different jobs to control which images we build in
our dsvm job.  This is needed so we can allow diskimage-builder more
control of which images to build.

Remove hardcoded apt-get for installing openssh-server, we can now use
the openssh-server element from diskimage-builder (which supports more
operating systems).

Change-Id: Ib6f48648e9ada8fd626a1811cdce262df0768c08
Signed-off-by: Paul Belanger <pabelanger@redhat.com>