Updated openstack/openstack Project: openstack-infra/nodepool  deb0aecc80f5e0a349fdc4fad1aafcb31161146d
Remove script-dir

Script-dir was only used for snapshot building.  Since that is no
longer supported, remove it.

It was also passed through to DIB as an evironment variable.
However, the OpenStack infra element that used it defaults to
a sensible value.  Users who still want to use that can add it
as an environment variable to appropriate diskimage entries.

Change-Id: Ie72bb9675fdad6364589a6e3823f4ef9a0228e86