Updated openstack/openstack Project: openstack/ironic  d8b04680d4491235cc395ec9f0743266db11f031
Adds another validation step when using dynamic allocation

The OneView driver currently allows nodes to use Server Profile
Templates (SPTs) that are marked to use Virtual MACs. This is an issue
because the MAC the machine is going to use will only be specified at
profile application. The driver only works with physical MACs in the
dynamic allocation case. This should be checked in all interfaces'
validations.

Change-Id: I5b779ac58ce096a52876bace77702e828aaac711
Depends-On: I21a1ebfc14ec33c661a1e7c7c1dbdcdcab04f119
Closes-Bug: 1627818
Co-Authored-By: Charlle Daniel <charlledaniel@lsd.ufcg.edu.br>