Updated openstack/openstack Project: openstack/trove  2a42adcf85e54aaab407d54e43a449c97e59cce6
Have inst-error delete before inst-create-wait

Before we wait for the the instance-create stuff to finish,
make sure the error instances are deleted.   That will free
up resources faster and hopefully make the gate run quicker.

Do the same thing with the module- and cfggrp- instances
(before the backup ones are created).

Change-Id: I6c70975538f1e2f656c31cfaf63fc69e2be821c0
Depends-On: I3bbe3bafa7ea3e627272103ac16a38f6a32a8a06