Updated openstack/openstack Project: openstack-dev/hacking  5917e798d7bbd35e0eea27c2cdfb314af778280e
Add Constraints support

Adding constraints support to libraries is slightly more complex than
services as the libraries themselves are listed in upper-constraints.txt
which leads to errors that you can't install a specific version and a
constrained version.

This change adds constraints support by also adding a helper script to
edit the constraints to remove hacking.

Change-Id: Id01984f5a986cf84a8fff3f028bb6246364ea2ff