Updated openstack/openstack Project: openstack/debtcollector  00d2db18a99cbc57feab5b91e6d4165f6f5de3c5
Add Constraints support

Adding constraints support to libraries is slightly more complex than
services as the libraries themselves are listed in upper-constraints.txt
which leads to errors that you can't install a specific version and a
constrained version.

This change adds constraints support by also adding a helper script to
edit the constraints to remove debtcollector.

Change-Id: Ie1e79f7bd0cb2a6b20d0485e12cd05522c2a0f9f