Updated openstack/openstack Project: openstack-infra/project-config  df67e4006f558a89dee8991bd38afe0c4b1948cc
move release step for updating constraints to after the PyPI upload

We have a race condition in the jobs for patches to
openstack/requirements submitted by the release jobs because the
packages are not yet on PyPI when the patch is proposed to update the
constraints file. This change moves the constraint update portion of the
release job to a separate job and schedules it to run as a dependency of
the PyPI upload job.

Change-Id: Idce73ca3813858050e8f64ece44e838266aaf5d6
Signed-off-by: Doug Hellmann <doug@doughellmann.com>