Updated openstack/openstack Project: openstack/nova  7e1f1f68ebc05deabcb308d5918c88563b483833
Mark sibling CPUs as 'used' for cpu_thread_policy = 'isolated'

'isolated' CPU allocation thread policy is guarantee
that no vCPUs from other guests wouldn't be able to be
placed on the cores of booted VM (In this case core is
a set of sibling vCPUs).

But we still able to boot VMs with 'dedicated' CPU
allocation policy on these cores. This problem is actual
for hosts without HyperThreading. In this case sets of
siblings vCPUs are empty for each core but we are still
trying to work with them as with HyperThreading cores.
This causes the problem when one "isolated" core
is used by several VMs.

To fix it we must use method unpin_cpus_with_siblings()
only if NUMA cell has siblings (i.e. has HyperThreading).
For cells without HyperThreading CPU isolation is
guaranteed by 'dedicated' CPU allocation policy.

Closes-Bug: #1635674

Change-Id: I8f72187153c930cd941b7ee7e835a20ed0c0de03