Updated openstack/openstack Project: openstack/neutron  96587ce6afc4d605310a06da1b5f3b21601aecb1
Calculate IPv4 DHCP subnets once for metadata

When enabling metadata, we iterate through the subnets
on a network multiple times.  Do it only once at the
beginning and return early if there are no candidates.
Follow-on to comments in an earlier review,
https://review.openstack.org/#/c/293237

Had to fix a few tests that were creating "fake" subnets
without an ip_version attribute or passing a network
mock instead of a fake one.

Change-Id: I57dfeec339a072e78242373bf793dbbf04e8e4c3