Updated openstack/openstack Project: openstack-infra/shade  7b17ccb0035e33707736b5d6d4ceb9d77fa74975
Make munch aware assertEqual test method

Many of our assertEqual calls use compare munches. When there is a
problem, the printout of the munch is all on one line, which makes
reading it hard. Make an assertEqual in our base class which turns
munches into dicts if they're passed in.

Change-Id: I478548c5146cd6f86b8957f527ccd3d78268156f