Updated openstack/openstack Project: openstack-infra/tripleo-ci  1cade7c8d052f57ea786e6a9538796a0ad2fb315
Add --setup-nodepool-files option

This option will setup the expected nodepool files that are not present
when using multinode deployments on systems not spawned by nodepool. We
don't have to use this option in tripleo-ci since nodepool spawned
instances are in use. However, this is a useful helper option when doing
local deployments.

Change-Id: I8fa2f21fb693307440f661606e3193e1aa942c0d