Updated openstack/openstack Project: openstack-infra/tripleo-ci  a0faab2357e58b4c2acce38a6521f04d226db526
Add 3 node multinode job

The 3 node multinode job tests with a primary node and 2 subnodes. The
primary node is used for the undercloud. The 2 subnodes are used for the
overcloud. The overcloud services are split between pacemaker managed
services and non pacemaker managed services on the 2 subnodes. The first
subnode has the minimal set of services for an all-in-one overcloud with
the exception of any pacemaker service (galera, rabbitmq, etc) which are
on the second subnode.

This job adds coverage of non-default composable roles since a custom
roles data yaml file has to be used to split up the services in this
manner across the 2 subnodes.

Depends-On: I8fc39e6d18cd70ff881e2a284234b26261018d67
Change-Id: I0cee6ff8b5c3002d45a6e66e4657baffa99ac5a6