Updated openstack/openstack Project: openstack/governance  d57494859b8721d81f0448cda8b556342cbc941f
Amend reference/PTI with supported distros

It became clear that Ubuntu and CentOS are the 2 most-popular platforms
when deploying OpenStack.

https://www.openstack.org/assets/survey/April-2016-User-Survey-Report.pdf

We recently saw some cases where some projects implemented some features
that could not work on these platforms.

Example: https://review.openstack.org/#/c/401985/
- A feature in Neutron couldn't work with latest CentOS.
- Neutron and more projects are not gating on CentOS platform, only
  Ubuntu.

The goal of this change is to document that we would like to support
Ubuntu and CentOS in our testing and also collaborate with distros when
a dependency is not fresh enough to bring new features in OpenStack.

Change-Id: I8d63a883b29eda2757656008be3c6af83b8f9e88