Updated openstack/openstack Project: openstack/cinder  732d980b2ca1ec7d9dc95a5bed74dd5858b32394
Make workers ORM compatible with SQLAlchemy 1.1.4

We don't currently support SQLAlchemy 1.1.4, but in preparation of the
day we do we can start making simple changes that will facilitate the
transition.

In SQLAlchemy 1.1 the .autoincrement directive is no longer implicitly
enabled for a composite primary key column, so the id field in the
workers ORM class will no longer be set as autoincrement, which means
that when creating a worker we will not be updating the id field of the
instance and it will left as None in memory while it is properly set in
the DB.

This patch resolves this issue by defining the id field as an auto
increment field explicitly instead of relying on the implicit definition
exiting in SQLAlchemy 1.0.

Closes-bug: #1651448
Change-Id: I570a4f0ea7850478097a735631031c75f0a65c5a