Updated openstack/openstack Project: openstack/keystone  8a5a5167dbaeacd46a329191968258965853b0cc
Invalidate token cache after token delete

After deleting a token, the token cache needs to be invalidated.
This was only happenning for persistent tokens.

This patch also calls the token cache invalidation on the token provider
so that it happens regardless of the token provider being used or if the
token has actually been persisted.

Change-Id: I496531a30559f0cf021c4478404093a99f2fbe3d