Updated openstack/openstack Project: openstack/openstackdocstheme  edec81e7400582730a41255c1b067e7b8d514a13
Add Constraints support

Adding constraints support to libraries is slightly more complex than
services as the libraries themselves are listed in upper-constraints.txt
which leads to errors that you can't install a specific version and a
constrained version.

This change adds constraints support by also adding a helper script to
edit the constraints to remove openstackdocstheme.

Change-Id: I4ed268e3fec3754a4344c57509ff6969406335ad