Updated openstack/openstack Project: openstack/python-keystoneclient  d12de79636d4584e48224576d4b81a23d0fe6d4e
Add Constraints support

Adding constraints support to libraries is slightly more complex than
services as the libraries themselves are listed in upper-constraints.txt
which leads to errors that you can't install a specific version and a
constrained version.

This change adds constraints support by also adding a helper script to
edit the constraints to remove python-keystoneclient.

Change-Id: I8933945ab8948c43022963ba6cff4d38cec5a04f