Updated openstack/openstack Project: openstack/python-saharaclient  fb803225f3e7c48958acc86a9e9b747d598320ee
Add Constraints support

Adding constraints support to libraries is slightly more complex than
services as the libraries themselves are listed in upper-constraints.txt
which leads to errors that you can't install a specific version and a
constrained version.

This change adds constraints support by also adding a helper script to
edit the constraints to remove python-saharaclient.

Change-Id: Ie448db6f78865a3dddadf7dc659c7546a5e0fa72