Updated openstack/openstack Project: openstack/stevedore  79f75a7f7001e2e0ed6228ed819a2476b97c7b13
Add Constraints support

Adding constraints support to libraries is slightly more complex than
services as the libraries themselves are listed in upper-constraints.txt
which leads to errors that you can't install a specific version and a
constrained version.

This change adds constraints support by also adding a helper script to
edit the constraints to remove stevedore.

Change-Id: I7a1285a2c3882b02987bfa95950689f5282aa407