Updated openstack/openstack Project: openstack/heat  77d94c2312a93e407aed885da90e6985babbc101
Update the constraint value of schema 'ACCESS_TYPE' for Manila share

Update the constaint value to make it up to date. Manila support 4
kinds of access types wich are 'ip', 'user', 'cert', 'cephx', and,
'cephx' is the newer type in Manila which is supported on version '2.13'.
So update the version of Manila client to '2' incidentally.

Change-Id: I230f2bc7268e9f25f77d7b22f7a68b6ca37f6d2a
Closes-Bug: #1649217