Updated openstack/openstack Project: openstack/cinder  c6262925c334864b080b33d494c3077a7243fad7
Fix VMAX unit test timeout issue

The unit test test_delete_volume_fast_success timed out because
it is doing 60 retries of 10 sec intervals. This patch set
the number of retries and interval to 0 in the test. Before this
fix, this unit test hung on my setup. After this fix, this unit
test finished in 3 secs and the total of 297 VMAX unit tests
finished in 11 secs on my setup.

Change-Id: I4edfff4c3127623998f251ff014c6f3c11bce4b2
Closes-Bug: #1651615