Updated openstack/openstack Project: openstack/nova  5c9d96a3d49b433d7feb1de9e4cb28056c900c71
conf: Deprecate more nova-net options

They're multiplying!

Or not. A few nova-net options were removed in I4a6d534c9. Remove some
that we missed, updating the release note in the process.

Note that the 'auto_assign_floating_ip' is only used in the FloatingIP
mixin, which in turn is only used with various nova-network managers.
This means the option is dependent on the 'use_neutron' configuration
option like mostly ever other option in the file. As such, the help text
for that option is updated accordingly.

Change-Id: If396a7b10562e2fa43cdf4b09e4632cc6e57391a
Implements: bp centralize-config-options-ocata