Updated openstack/openstack Project: openstack-infra/tripleo-ci  b778715e74c0d0578081a7fb5d653486fda76c63
Echo to deploy.env the stable release for the UC upgrade

The stable_release variable needs to be updated in the
deploy.env, because this file is sourced on each tripleo.sh
call. If we do not update the stable_release variables the
major upgrade for Undercloud is doing simply a minor
newton -> newton.

Change-Id: Ibc6fc34e55455c4d06ccc8bc717eb7c3fe8d499f