Updated openstack/openstack Project: openstack-infra/project-config  5360d4845d25bf3715ec954768090890c2cd39e3
Align the dsvm-neutron-full-ssh job

The tempest-dsvm-neutron-full-ssh diverged slightly from the
tempest-dsvm-neutron-full job over time (TLS is missing).
The target is to have the -ssh job eventually replace the
non ssh one, so we need them to be setup in the same way
(apart from the ssh setting) to collect valid data.

Change-Id: I636371982645cc1f3cd1ff381c3c0707c654f96a