Updated openstack/openstack Project: openstack/heat-translator  6e6a02e1b58934de6e01edd76ae3c5e60dc27a7c
Adding conf files as package data

I've been packaging this python module and found that setup.py was not
including the conf files in translator/conf/ and could not use it
without these files to generate the tacker.conf file with oslo gen
config.

Adding those files to the setup.py file so that a setup.py install will
include them when the installation is done.

Change-Id: Ifda75a9167aaceb51dfafe32f28ee47d6f8a4e0b
Signed-off-by: Dan Radez <dradez@redhat.com>