Updated openstack/openstack Project: openstack/python-keystoneclient-kerberos  59b7e97e2da962714fddebc8b5cafae7365816f2
retire keystoneclient-kerberos

The keystoneclient-kerberos project is being retired. The project
was seldomly used, untested, and any logic it had was moved to
the keystoneauth project.

Change layout.yaml and projects.yaml according to the retirement
instructions [1].

[1] http://docs.openstack.org/infra/manual/drivers.html#retiring-a-project

See related commits:

  - infra related patches to retire the repo
    I7a6c0ed3f799904815d928f2572fec657d3e2b10

  - keystoneclient-kerberos deprecation warning
    I9877a044f98bcb3479bd2d853aedd2d2a5b10b3b

  - requirements update
    Ib1307a21a4a8697dd28fb01ba5b2df6850cc4f3a

  - mailing list announce
    http://lists.openstack.org/pipermail/openstack-dev/2016-September/104721.html

  - codesearch (showing no impact on consuming projects)
    http://codesearch.openstack.org/?q=keystoneclient-kerberos&i=nope&files=&repos=

This can only merge after the following change has
merged in project-config:
https://review.openstack.org/#/c/413790/
Not using a formal depencency since the Change-Id of the project-config
is reused in an abandoned change.

Change-Id: I2b31d7c07a7456b4208d928ca2c013ccea88874d