Updated openstack/openstack Project: openstack/python-zaqarclient  3c8b1a28e009f9ecaefee6d2511d2c2f0ff52154
Add Constraints support

Adding constraints support to libraries is slightly more complex than
services as the libraries themselves are listed in upper-constraints.txt
which leads to errors that you can't install a specific version and a
constrained version.

This change adds constraints support by also adding a helper script to
edit the constraints to remove python-zaqarclient.

Change-Id: I43da802224be8e3ff249a8fa1d5bd4d6475ba391