Updated openstack/openstack Project: openstack/neutron  6321ba7bed6ad4b0aa8ae6891d50265ac229071f
Change passing session to context for TypeDriver

We have a number of functions that expect to get session as one
of the arguments. Passing session is not correct and this prevents
of using new enginefacade which expects context to be passed and
session will be injected in current context.

NeutronLibImpact

Partially-Implements blueprint: enginefacade-switch

Change-Id: I68ac442f11a6698bc43210fae877ba3809b1c02d