Updated openstack/openstack Project: openstack/cinder  ad38bc35720b3a2496c752791fd6a88f8bcb6ae4
Update is_up from Service OVO class to match ORM

Our services' update_at field is also being updated now on creation [1],
so there's no longer a need to check for the created_at field for the
heartbeat in is_up method.

Since this change was introduced in Newton it is now safe for this patch
to remove the unnecessary check on created_at in method is_up of the
Service OVO class.

[1] https://review.openstack.org/#/c/318572/22/cinder/db/sqlalchemy/models.py@73

Trivialfix

Change-Id: Id0e30656901810128415fd9544c28cafb22c6592