Updated openstack/openstack Project: openstack-infra/shade  8e115beb1214365e42a074b0ea6e232216a1696a
Combine list of calls with list of request assertions

It's really awkward to assert that the request to call 14 had a json
body that looks like something. Instead, it's more natural to say
"there's going to be a POST call to $url and it should have this body"

Make the assert_calls helper method also look for optional json and
headers content.

Change-Id: I12a50d46416c80b052cf30b605b2820899a1eb77