Updated openstack/openstack Project: openstack-infra/shade  9fbacfa2f3455d53b138cb159b04e831c5199516
Replace mocks of swiftclient with request_mock

The string handling around headers and swiftclient seems a little off.
Note the b'' we had to add to the public acl as well as the str() in
get_container_access and the str() around the header in test_object. The
swift docs say that headers should be unicode strings - but in 3.5 we
get bytes in some places that it seems like we should be getting
strings. The b'' and str() casting make the tests work - but I'm not
convinced that's the correct thing to do in these cases.

Change-Id: I2d7c77eaab690f578eaba80490145bc16a23222c