Updated openstack/openstack Project: openstack-infra/shade  35f40bc43840fe253d68f7c34bcbd37f77ab6897
Stop using full_listing in prep for REST calls

The full_listing parameter to swiftclient causes it to do a
batching/paging approach all the time. We can do it smarter, but for
now, just remove it so that we can do an equal logic REST conversion.
Then we'll add paging support smartly to the REST calls and always do
it.

Change-Id: I743871c5f87fba28d16b38fe23c51fd0cdd7eecb