Updated openstack/openstack Project: openstack/requirements  46137f70ea54b6ca40465607e33ba514d20da04f
Raise taskflow to >= 2.7.0

Many fixes have gone in to the taskflow library, but
we have not raised the minimum requirement for quite
some time.

I9658fbac97e4074b551a50d6016a92806a3636c1 added a new
way to set the logging fail_formatter. In order to use
this method, we need the minimum version raised to the
version where it was released.

Change-Id: I7e4a234d4c26363cb4bcc463597d57913f81740d